Lower Go version to 1.20 for plugin
The build system for the kinesis plugin is rather complex - in short, we
pull this code into an AWS AL2 based container where Go is installed,
compile the plugin and then copy it out for inclusion in our own image.
This means that the minimum Go version specified in this code must be
no higher than the Go version available in the AWS AL2 image (which is
currently Go 1.20). Unfortunately this version is no longer receiving
reliability or security updates, but we'll have to deal with this issue
in the future.